requirements.txt
streamlit
alpaca-py
pandas
plotly